robot can now follow all line types

+++ RobotCode/src/erhs53/utilities/MathUtils.java
@@ -40,7 +40,7 @@ public class MathUtils {
 		return sum / data.length;
 	}
 	
-	public static double map(double x, double in_min, double in_max, double out_min, double out_max) {
+	public static float map(float x, float in_min, float in_max, float out_min, float out_max) {
 	  return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 	}
